Assign adoptedStyleSheets rather than push
Some older browsers such as Chrome 99 have `adoptedStyleSheets` but it
is a frozen array, rather than an `ObservableArray`, and can only be
assigned to rather than `push()`ed to. Pushing to a frozen array causes
an error to be thrown and so the polyfill doesn't work in those browsers
if we continue to use push.

The fix for this is to always use an assignment on the array, it's a
little slower the impact will be negligible.

See https://chromestatus.com/feature/5638996492288000 for some more
detail.